requirements.txt
boto3==1.34.148
fastapi==0.112.0
python-dotenv==1.0.1
uvicorn==0.30.1
httpx==0.27.0
Pillow==10.4.0
Pygments==2.18.0
requests==2.32.3
pytest==8.3.2
python-multipart